chore(tools): remove unused scripts
